requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-file-share==12.3.0
azure-functions
requests
aliyun-log-python-sdk
aiohttp